Move jujud agent config out to agent package
