+++ cmd/jujud/agent/caasunitinit.go
@@ -20,10 +20,10 @@ import (
 	"github.com/juju/loggo"
 	"github.com/juju/names/v4"
 
+	"github.com/juju/juju/agent/config"
 	"github.com/juju/juju/agent/tools"
 	"github.com/juju/juju/caas"
 	jujucmd "github.com/juju/juju/cmd"
-	"github.com/juju/juju/cmd/jujud/agent/config"
 	"github.com/juju/juju/juju/sockets"
 	"github.com/juju/juju/worker/uniter"
 	"github.com/juju/juju/worker/uniter/runner/jujuc"
