+++ cmd/jujud/introspect/introspect_test.go
@@ -24,7 +24,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/cmd/jujud/agent/config"
+	"github.com/juju/juju/agent/config"
 	"github.com/juju/juju/cmd/jujud/introspect"
 	"github.com/juju/juju/testing"
 )
