+++ cmd/jujud/main.go
@@ -26,12 +26,12 @@ import (
 	"github.com/juju/version/v2"
 
 	"github.com/juju/juju/agent/addons"
+	"github.com/juju/juju/agent/config"
 	k8sexec "github.com/juju/juju/caas/kubernetes/provider/exec"
 	jujucmd "github.com/juju/juju/cmd"
 	agentcmd "github.com/juju/juju/cmd/jujud/agent"
 	"github.com/juju/juju/cmd/jujud/agent/agentconf"
 	"github.com/juju/juju/cmd/jujud/agent/caasoperator"
-	"github.com/juju/juju/cmd/jujud/agent/config"
 	"github.com/juju/juju/cmd/jujud/dumplogs"
 	"github.com/juju/juju/cmd/jujud/introspect"
 	"github.com/juju/juju/cmd/jujud/run"
