+++ cmd/jujud/run/run.go
@@ -23,9 +23,9 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/juju/juju/agent"
+	"github.com/juju/juju/agent/config"
 	"github.com/juju/juju/caas"
 	jujucmd "github.com/juju/juju/cmd"
-	"github.com/juju/juju/cmd/jujud/agent/config"
 	"github.com/juju/juju/core/machinelock"
 	jujuos "github.com/juju/juju/core/os"
 	"github.com/juju/juju/core/paths"
