+++ cmd/jujud/run/run_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/juju/utils/v3/exec"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/cmd/jujud/agent/config"
+	"github.com/juju/juju/agent/config"
 	"github.com/juju/juju/core/machinelock"
 	"github.com/juju/juju/juju/sockets"
 	"github.com/juju/juju/testing"
